components/e2e/pyproject.toml
[project]
name = "pinta-e2e"
version = "0.0.0"
description = "End to end tests for elevation production system"
readme = "README.md"
requires-python = ">=3.12"


[project.urls]
Homepage = "https://github.com/nlsfi/pinta"
Repository = "https://github.com/nlsfi/pinta"

[build-system]
requires = ["uv_build>=0.10.9,<0.12.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-name = "pinta_e2e_utils"
include = ["py.typed"]

[tool.uv]
default-groups = ["dev", "lint"]

[tool.uv.sources]
pinta-qgis-plugin = { workspace = true }
pinta-lib = { workspace = true }

[dependency-groups]
dev = [
  "pinta-lib",
  "pinta-qgis-plugin",
  "pytest>=9.0.2",
  "pytest-qgis>=3.0.0",
  "pytest-xdist>=3.8.0",
]
lint = [
  "prek",
  "mypy",
  "ruff",
]

[tool.pytest.ini_options]
testpaths = [
  "test",
]
addopts = "-n auto --import-mode=importlib"
